Fix scheduling logs.

+++ application/scheduler/run_jobs.py
@@ -4,7 +4,6 @@ import json
 import requests
 import os
 from datetime import datetime
-from pytz import timezone
 
 # TODO The port should not be hard-coded, it should be determined through some sort of Daudit config
 DAUDIT_PORT = 3000
@@ -19,10 +18,8 @@ def main():
             config_json = json.load(config_file)
             print("Sending the following jobs:")
             list_of_jobs = []
-            tz = 'US/Eastern'
-            hour = datetime.now(timezone(tz)).hour
+            hour = datetime.now().hour
             for (key, value) in config_json.items():
-                # TODO: Make it so that we check if a job should be run based on last ran
                 if value["hour_of_day"] == hour:
                     list_of_jobs.append({
                         "id": key,
@@ -30,23 +27,23 @@ def main():
                         "channel_id": value['channel_id']
                     })
                 value['last_ran'] = datetime.now().strftime("%d/%m/%Y %H:%M")
-                print("\t%s, %s" % (key, value['channel_id']))
 
-            # Send the jobs
-            print("Current time (%s): %s" % (datetime.now(), tz))
-            print("Sending: " + str(list_of_jobs))
-            r = requests.post(url=DAUDIT_URL, json=list_of_jobs)
+            # Send the jobs only if there are any to send
+            if len(list_of_jobs) > 0:
+                print("Current time: %s" % datetime.now())
+                print("Sending: " + str(list_of_jobs))
+                r = requests.post(url=DAUDIT_URL, json=list_of_jobs)
 
-            print("Response from Daudit: " + str(r.status_code))
-            if r.status_code == 200:
-                print("Jobs successfully posted to run on Daudit!")
-            else:
-                print("ERROR: jobs not successfully posted on Daudit!")
+                print("Response from Daudit: " + str(r.status_code))
+                if r.status_code == 200:
+                    print("Jobs successfully posted to run on Daudit!")
+                else:
+                    print("ERROR: jobs not successfully posted on Daudit!")
 
-            # Update last ran field in file
-            config_file.seek(0)
-            config_file.write(json.dumps(config_json))
-            config_file.truncate()
+                # Update last ran field in file
+                config_file.seek(0)
+                config_file.write(json.dumps(config_json))
+                config_file.truncate()
     else:
         print("Config file does not exist. Please schedule a job first.")
